+++ modules/core/src/main/java/org/apache/ignite/IgniteIllegalStateException.java
