+++ modules/core/src/main/java/org/apache/ignite/IgniteInterruptedException.java
