+++ modules/core/src/main/java/org/apache/ignite/internal/IgniteClientDisconnectedCheckedException.java
